requirements.txt
MetaTrader5
discord.py